requirements.txt
elasticsearch>=8.12.0
sentence-transformers>=2.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.30.0
langchain>=0.1.0
langchain-elasticsearch>=0.1.0
requests>=2.31.0
tqdm>=4.66.0
